requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
pkg-resources==0.0.0
python-decouple==3.4
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
